Break ThemeContext component to partials
