+++ modules/react-components/src/theme/theme.tsx
@@ -16,286 +16,19 @@
  * under the License.
  */
 
-import React, { createContext, Dispatch, useReducer } from "react"; 
-import { Theme, Themes } from "@wso2is/theme";
-
-/*
- * Themes index less file.
- */
-const ThemeLess = (theme) => { return `themes-less/themes/${theme}/index.less` };
-
-/*
- * Theme options type.
- */
-const themes = Themes as Themes; // Convert Themes array to string literal
-type ThemeTypes = typeof themes[number];
-
-/*
- * Theme compile options interface.
- */
-interface ThemeCompileOptionsInterface {
-    "@primaryColor"?: string;
-    "@pageBackground"?: string;
-}
-
-/*
- * ThemeContext state interface.
- */
-interface ThemeContextState {
-    appName: string;
-    copyrightText: string;
-    css: string;
-    logo: string;
-    productName: string;
-    styles: ThemeCompileOptionsInterface;
-    theme: ThemeTypes;
-}
-
-/*
- * ThemeContext interface.
- */
-interface ThemeContextInterface {
-    dispatch: ({ type }: {type: string}) => void;
-    compile: (options: ThemeCompileOptionsInterface) => void;
-    setAppName: (name: string) => void;
-    setCopyrightText: (text: string) => void;
-    setCSS: (css: string) => void;
-    setLogo: (url: string) => void;
-    setProductName: (name: string) => void;
-    setStyles: (styles: ThemeCompileOptionsInterface) => void;
-    setTheme: (theme: ThemeTypes) => void;
-    state: ThemeContextState;
-}
-
-/**
- * Initial theme context state.
- */
-const themeInitialState: ThemeContextState = {
-    appName: "",
-    copyrightText: "",
-    css: "",
-    logo: "",
-    productName: "",
-    styles: {},
-    theme: "default"
-}
-
-/**
- * ThemeContext reducer actions.
- */
-const themeContextReducerActions = {
-    SET_APP_NAME: "SET_APP_NAME",
-    SET_COPYRIGHT_TEXT: "SET_COPYRIGHT_TEXT",
-    SET_CSS: "SET_CSS",
-    SET_LOGO_URL: "SET_LOGO_URL",
-    SET_PRODUCT_NAME: "SET_PRODUCT_NAME",
-    SET_STYLES: "SET_PRODUCT_NAME",
-    SET_THEME: "SET_THEME"
-};
-
-/**
- * Set Theme CSS reducer action.
- *
- * @param {string} css - Compiled CSS string.
- */
-const setCSS = (css: string) => ({
-    payload: css,
-    type: themeContextReducerActions.SET_CSS
-});
-
-/**
- * Set User Defined LESS variable values.
- *
- * @param {ThemeCompileOptionsInterface} styles - LESS ModifyVars
- */
-const setStyles = (styles: ThemeCompileOptionsInterface) => ({
-    payload: styles,
-    type: themeContextReducerActions.SET_STYLES
-});
-
-/**
- * Set Application name reducer action.
- *
- * @param {string} name - Application name.
- */
-const setAppName = (name: string) => ({
-    payload: name,
-    type: themeContextReducerActions.SET_APP_NAME
-});
-
-/**
- * Set Application copyright text reducer action.
- *
- * @param {string} text - Application copyright text.
- */
-const setCopyrightText = (text: string) => ({
-    payload: text,
-    type: themeContextReducerActions.SET_COPYRIGHT_TEXT
-});
-
-/**
- * Set Product logo reducer action.
- *
- * @param {string} url - Url of logo image
- */
-const setLogo = (url: string) => ({
-    payload: url,
-    type: themeContextReducerActions.SET_LOGO_URL
-});
-
-/**
- * Set Product Name reducer action.
- *
- * @param {string} name - Product name.
- */
-const setProductName = (name: string) => ({
-    payload: name,
-    type: themeContextReducerActions.SET_PRODUCT_NAME
-});
-
-/**
- * Set Theme name reducer action.
- *
- * @param {ThemeTypes} theme - Theme. E.g. "default", "dark"
- */
-const setTheme = (theme: ThemeTypes) => ({
-    payload: theme,
-    type: themeContextReducerActions.SET_THEME
-});
-
-/**
- * ThemeContext reducer.
- *
- * @param {themeInitialState} [ state ] - Initial Theme Context State.
- * @param {any} action - Reducer actions.
- * @returns
- */
-const themeContextReducer = (state = themeInitialState, action) => {
-    switch (action.type) {
-        case themeContextReducerActions.SET_APP_NAME:
-            return {
-                ...state,
-                appName: action.payload
-            };
-        case themeContextReducerActions.SET_COPYRIGHT_TEXT:
-            return {
-                ...state,
-                copyrightText: action.payload
-            };
-        case themeContextReducerActions.SET_STYLES:
-            return {
-                ...state,
-                styles: action.payload
-            };
-        case themeContextReducerActions.SET_LOGO_URL:
-            return {
-                ...state,
-                logo: action.payload
-            };
-        case themeContextReducerActions.SET_PRODUCT_NAME:
-            return {
-                ...state,
-                productName: action.payload
-            };
-        case themeContextReducerActions.SET_THEME:
-            return {
-                ...state,
-                theme: action.payload
-            };
-        case themeContextReducerActions.SET_CSS:
-            return {
-                ...state,
-                css: action.payload
-            };
-        default:
-            return state;
-    }
-};
-
-/**
- * Theme compile method.
- * 
- * @param {any} dispatch - `dispatch` function from ReactContext.
- * @param {any} state - ThemeContext state.
- * @param {ThemeCompileOptionsInterface} options - Less copile options.
- */
-const handleCompileTheme = (dispatch, state, options: ThemeCompileOptionsInterface) => {
-    Theme.compile(ThemeLess(state.theme), { modifyVars: options })
-        .then((styles) => {
-            dispatch(setStyles(options));
-            dispatch(setCSS(styles));
-        });
-};
-
-/**
- * Set Application Name method.
- * 
- * @param {any} dispatch - `dispatch` function from ReactContext.
- * @param {string} name - Application name.
- */
-const handleSetAppName = (dispatch, name: string) => {
-    dispatch(setAppName(name));
-};
-
-/**
- * Set CSS string method.
- *
- * @param {any} dispatch - `dispatch` function from ReactContext.
- * @param {string} css - Compiled CSS string.
- */
-const handleCSS = (dispatch, css: string) => {
-    dispatch(setCSS(css));
-};
-
-/**
- * Set Copyright Text method.
- * 
- * @param {any} dispatch - `dispatch` function from ReactContext.
- * @param {string} text - Copyright text.
- */
-const handleSetCopyrightText = (dispatch, text: string) => {
-    dispatch(setCopyrightText(text));
-};
-
-/**
- * Set Logo URL method.
- * 
- * @param {any} dispatch - `dispatch` function from ReactContext.
- * @param {string} url - Product logo URL.
- */
-const handleSetLogo = (dispatch, url: string) => {
-    dispatch(setLogo(url));
-};
-
-/**
- * Set Product Name method.
- * 
- * @param {any} dispatch - `dispatch` function from ReactContext.
- * @param {string} name - Product name.
- */
-const handleProductName = (dispatch, name: string) => {
-    dispatch(setProductName(name));
-};
-
-/**
- * Set LESS style configuration method.
- * 
- * @param {any} dispatch - `dispatch` function from ReactContext.
- * @param {string} styles - LESS ModifyVars object.
- */
-const handleStyles = (dispatch, styles: ThemeCompileOptionsInterface) => {
-    dispatch(setStyles(styles));
-};
-
-/**
- * Theme toggle method.
- * 
- * @param {any} dispatch - `dispatch` function from ReactContext.
- * @param {ThemeTypes} theme - Theme. E.g. "default", "dark"
- */
-const handleThemeToggle = (dispatch, theme: ThemeTypes) => {
-    dispatch(setTheme(theme));
-};
+import {
+    handleCompileTheme,
+    handleCSS,
+    handleProductName,
+    handleSetAppName,
+    handleSetCopyrightText,
+    handleSetLogo,
+    handleStyles,
+    handleThemeToggle
+} from "./actions";
+import React, { createContext, Dispatch, useReducer } from "react";
+import { ThemeCompileOptionsInterface, ThemeContextInterface, ThemeTypes } from "./models";
+import { themeContextReducer, themeInitialState } from "./reducer";
 
 /**
  * ThemeContext.
