+++ modules/theme/src/index.js
@@ -19,14 +19,9 @@
 
 export * from "./theme";
 
-/*
- * Export default theme variables
- */
+export const ThemeLessIndex = (theme) => { return `themes-less/themes/${theme}/index.less`; };
 export const defaultThemeVariables = require("../dist/theme-variables.json");
-export const Themes = [
-    "dark",
-    "default"
-];
+export const Themes = [ "dark", "default" ];
 
 // Icons
 export const AlertIcon = require("../dist/lib/themes/default/assets/images/icons/alert-icon.svg");
